Don't push 64 bits and pop 32 bits on 32 bit systems

+++ more/viewport.c
@@ -1013,7 +1013,7 @@ MoreDrawStatusLine(
                       FirstViewportLine,
                       LastViewportLine,
                       TotalLines,
-                      LastViewportLine * 100 / TotalLines,
+                      (DWORD)(LastViewportLine * 100 / TotalLines),
                       &MoreContext->SearchString);
     } else {
         YoriLibYPrintf(&LineToDisplay,
@@ -1022,7 +1022,7 @@ MoreDrawStatusLine(
                       FirstViewportLine,
                       LastViewportLine,
                       TotalLines,
-                      LastViewportLine * 100 / TotalLines);
+                      (DWORD)(LastViewportLine * 100 / TotalLines));
     }
 
     //
